+++ crypto/lib/src/hash_sink.dart
@@ -3,6 +3,7 @@
 // BSD-style license that can be found in the LICENSE file.
 
 import 'dart:typed_data';
+import 'dart:math' as math;
 
 import 'package:typed_data/typed_data.dart';
 
@@ -25,9 +26,10 @@ abstract class HashSink implements Sink<List<int>> {
   /// used across invocations of [_iterate].
   final Uint32List _currentChunk;
 
-  /// Messages with more than 2^64-1 bits are not supported.
-  /// So the maximum length in bytes is (2^64-1)/8.
-  static const _maxMessageLengthInBytes = 0x1fffffffffffffff;
+  /// Messages with more than 2^53-1 bits are not supported. (This is the
+  /// largest value that is representable on both JS and the Dart VM.)
+  /// So the maximum length in bytes is (2^53-1)/8.
+  static const _maxMessageLengthInBytes = 0x0003ffffffffffff;
 
   /// The length of the input data so far, in bytes.
   int _lengthInBytes = 0;
@@ -126,7 +128,7 @@ abstract class HashSink implements Sink<List<int>> {
 
     if (_lengthInBytes > _maxMessageLengthInBytes) {
       throw new UnsupportedError(
-          'Hashing is unsupported for messages with more than 2^64 bits.');
+          'Hashing is unsupported for messages with more than 2^53 bits.');
     }
 
     var lengthInBits = _lengthInBytes * bitsPerByte;